raspberry_pi/requirements.txt
opencv-python>=4.5.0
websocket-client>=1.0.0
numpy>=1.19.0
picamera2>=0.3.0; platform_machine == "armv7l" or platform_machine == "aarch64"

